move options editors from UI to core
