+++ public/app/core/components/OptionsUI/units.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 
 import { FieldConfigEditorProps, UnitFieldConfigSettings } from '@grafana/data';
-
-import { UnitPicker } from '../UnitPicker/UnitPicker';
+import { UnitPicker } from '@grafana/ui';
 
 export const UnitValueEditor: React.FC<FieldConfigEditorProps<string, UnitFieldConfigSettings>> = ({
   value,
